feat: PageHeader uses context
