+++ src/layouts/population/page/index.tsx
@@ -2,8 +2,6 @@
 import _ from 'lodash';
 import { formatShortDecimal, formatNumber, formatChangeNumber, formatChangePercent } from '../../../utils';
 import {
-  MainTitle,
-  SubTitle,
   Headline,
   ItemWrapper,
   CrossLink,
@@ -12,7 +10,6 @@ import {
 } from '../../../styles/MainContentStyles';
 import EntityTable from '../../../components/table/EntityTable';
 import EntityChart from '../../../components/chart/EntityChart';
-import PageHeader from '../../../components/PageHeader';
 import { Context } from '../../../utils/context';
 import ControlPanel from '../../../components/ControlPanel/ControlPanel';
 import { useContext } from 'react';
@@ -23,44 +20,10 @@ const PopulationPage = () => {
   const { clientData, clientAlias, tableData, clientProducts } = useContext(Context);
 
   const { LongName: prettyName } = clientData;
+
   const hasForecast = clientProducts => _.some(clientProducts, product => product.AppID === 3);
+
   const FormattedNumber = ({ number }) => <>{formatNumber(number)}</>;
-  const handleExport = async () => {
-    const IDReportRequest = {
-      FileName: `Population - ${prettyName}`,
-      Urls: [
-        {
-          Title: `Population - ${prettyName}`,
-          url: window.location.href,
-        },
-      ],
-      Action: 0,
-      EmailAddress: 'fabrice@id.com.au',
-    };
-
-    try {
-      const data = await postData(
-        'https://idreportserviceweb.azurewebsites.net/api/IDReportService/RequestReport/',
-        IDReportRequest,
-      ).then(res => {
-        console.log('Report Ok: ', res);
-      });
-      console.log(`Page report request: Population - ${prettyName}`);
-    } catch (error) {
-      console.error(error);
-    }
-  };
-
-  const postData = async (url = '', data = {}) => {
-    const response = await fetch(url, {
-      method: 'POST',
-      body: JSON.stringify(data),
-      headers: {
-        'Content-Type': 'application/json',
-      },
-    });
-    return response.json();
-  };
 
   const chartData = chartBuilder(tableData);
   const chartLineData = chartLineBuilder(tableData);
@@ -70,10 +33,6 @@ const PopulationPage = () => {
 
   return (
     <>
-      <PageHeader handleExport={handleExport}>
-        <MainTitle>{prettyName}</MainTitle>
-        <SubTitle>Population page</SubTitle>
-      </PageHeader>
       <Headline>
         The Estimated Resident Population of the {prettyName} was <FormattedNumber number={latestPop} /> as of the 30th
         June {latestYear}.
